Ayaengine/requirements.txt
absl-py
accelerate
aiofiles
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
async-timeout
asyncer
attrs
Babel
backoff
beautifulsoup4
bidict
bitsandbytes
bleach
cachetools
certifi
cffi
chainlit
chardet
charset-normalizer
chevron
ci-info
click
coloredlogs
comm
configobj
configparser
contourpy
cryptography
cycler
dataclasses-json
debugpy
decorator
deepdiff
defusedxml
Deprecated
deprecation
docker
effdet
emoji
etelemetry
exceptiongroup
executing
faiss-cpu
fastapi
fastjsonschema
fasttext
filelock
filetype
flatbuffers
fonttools
fqdn
frontend
frozenlist
fsspec
gast
google-api-core
google-auth
google-cloud-vision
google-pasta
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h5py
httpcore
httplib2
httpx
huggingface-hub
humanfriendly
idna
img2pdf
importlib_metadata
iopath
ipykernel
ipywidgets
isodate
isoduration
itsdangerous
jedi
Jinja2
joblib
json5
jsonpatch
jsonpath-python
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keras
kiwisolver
langchain
langchain-community
langchain-core
langchain-text-splitters
langdetect
langsmith
layoutparser
Lazify
libclang
literalai
looseversion
lxml
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mistune
ml-dtypes
mpmath
multidict
mypy-extensions
namex
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nibabel
nipype
nltk
notebook
notebook_shim
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
ocrmypdf
omegaconf
onnx==1.16.1
onnxruntime
opencv-python
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opt-einsum
optree
ordered-set
orjson
overrides
packaging
pandas
pandocfilters
parso
pathlib
pdf2image
pdfminer.six
pdfplumber
pexpect
pi_heif
pikepdf
pillow
pillow_heif
platformdirs
pluggy
portalocker
prometheus_client
prompt_toolkit
proto-plus
protobuf
prov
psutil
ptyprocess
pure-eval
pyasn1
pyasn1_modules
pybind11
pycocotools
pycparser
pydantic
pydantic_core
pydot
Pygments
PyJWT
PyMuPDF
PyMuPDFb
pyparsing
pypdf
pypdfium2
pytesseract
python-dateutil
python-dotenv
python-engineio
python-iso639
python-json-logger
python-magic
python-multipart
python-socketio
pytz
pyxnat
PyYAML
pyzmq
qtconsole
QtPy
rapidfuzz
rdflib
referencing
regex
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
Send2Trash
sentence-transformers
sentencepiece
simple-websocket
simplejson
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
sympy
syncer
tabulate
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
terminado
tf_keras
threadpoolctl
timm
tinycss2
tokenizers
tomli
torch
torchvision
tornado
tqdm
traitlets
traits
transformers
triton
types-python-dateutil
typing-inspect
typing_extensions
tzdata
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
uptrace
uri-template
urllib3
uvicorn
watchfiles
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
wrapt
wsproto
yarl
zipp
tesserocr
Cython
fastapi
uvicorn
easyocr